5-task commit

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 int c = 0;
-unsigned int nm = n ^ m;
+unsigned long int nm = n ^ m;
 while (nm != 0)
 {
 if (nm & 1)
